Convert loadMdx to ts
